findutils/grep.c: Add `-p` option.
This causes grep to only filter the input, but not signal matches
using the exit status. The rationale is that the lack of a match is
not automatically a sign of an error. This is especially problematic
with Bash's `set -eo pipefail`, consider this:

cat logfile | grep issue | sort -u

If you have no issues in the logs, grep will create one for you by
signaling an error here. The alternative is to use

cat logfile | (grep issue || [ $? -eq 1 ]) | sort -u

or to turn off error handling.